backend/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
flask-cors==6.0.0
supabase==2.15.2
pytest==8.3.5
reportlab==4.2.5
PyJWT==2.8.0
sendgrid==6.11.0 